+++ ddth-dao-core/src/main/java/com/github/ddth/dao/jdbc/annotations/ChecksumColumn.java
@@ -6,36 +6,24 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Annotation to mark DB table column list for primary key.
+ * Annotation to mark DB table's checksum column.
  * 
  * <p>
- * Usage - Single-column primary key:
+ * Usage:
  * </p>
  * 
  * <pre>
- * &#64;PrimaryKeyColumns(columns = { "id" })
+ * &#64;ChecksumColumn("checksum")
  * public class MyUserBoRowMapper extends AbstractRowMapper<UserBo> {
  *     ....
  * }
  * </pre>
- *
- * <p>
- * Usage - Multi-column primary key:
- * </p>
- * 
- * <pre>
- * &#64;PrimaryKeyColumns(columns = { "col1", "col2" })
- * public class MyUserBoRowMapper extends AbstractRowMapper<UserBo> {
- *     ....
- * }
- * </pre>
- * 
  * 
  * @author Thanh Nguyen <btnguyen2k@gmail.com>
- * @since 0.8.0
+ * @since 0.8.1
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.TYPE)
-public @interface PrimaryKeyColumns {
-    String[] columns();
+public @interface ChecksumColumn {
+    String value();
 }